Providing vertical slide gestures
