+++ lib/widgets/story_image.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
-import 'story_controller.dart';
-import 'utils.dart';
+import '../utils.dart';
+import '../controller/story_controller.dart';
 
 /// Utitlity to load image (gif, png, jpg, etc) media just once. Resource is
 /// cached to disk with default configurations of [DefaultCacheManager].
@@ -80,7 +80,7 @@ class StoryImage extends StatefulWidget {
   }) : super(key: key ?? UniqueKey());
 
   /// Use this shorthand to fetch images/gifs from the provided [url]
-  static StoryImage url(
+  factory StoryImage.url(
     String url, {
     StoryController controller,
     Map<String, dynamic> requestHeaders,
